+++ interpolation/adhier/config.go
@@ -3,18 +3,18 @@ package adhier
 // Config represents a configuration of the algorithm.
 type Config struct {
 	// The number of inputs.
-	Inputs uint16
+	Inputs uint
 	// The number of outputs.
-	Outputs uint16
+	Outputs uint
 	// The minimal level of interpolation. The nodes that belong to lower levels
 	// are unconditionally included in the surrogate.
-	MinLevel uint8
+	MinLevel uint32
 	// The maximal level of interpolation. The nodes that belong to this level
 	// are not refined, and, thus, the algorithm stops.
-	MaxLevel uint8
+	MaxLevel uint32
 	// The maximal number of nodes. The algorithm stops after reaching this many
 	// nodes.
-	MaxNodes uint32
+	MaxNodes uint
 	// The absolute error. The parameter is used for local refinement and is
 	// given in absolute units.
 	AbsError float64
@@ -23,11 +23,11 @@ type Config struct {
 	RelError float64
 	// The number of concurrent workers. The evaluation of the target function
 	// and the surrogate itself is distributed among this many goroutines.
-	Workers uint32
+	Workers uint
 }
 
 // DefaultConfig returns the default configuration of the algorithm.
-func DefaultConfig(inputs, outputs uint16) *Config {
+func DefaultConfig(inputs, outputs uint) *Config {
 	return &Config{
 		Inputs:   inputs,
 		Outputs:  outputs,
